Renamed governor to validator and fixed some naming conventions
